@@ -37,7 +37,7 @@
         {
             bool isContainer = item.CodeCreatorModels != null;
 
-            <div class="codecreator @if (isContainer) { @("codecreator-container"); }" data-id="@item.Identifier" data-type="@item.Type" title="@item.Type">
+            <div class="codecreator @if (isContainer) { @("codecreator-container"); }" data-id="@item.Identifier" data-type="@item.Type" data-customfactory="@item.CustomFactory" title="@item.Type">
 
                 @{string displayname = item.UserDisplayName != null ? item.UserDisplayName : item.DisplayName;}
 
@@ -134,7 +134,7 @@
 
         @foreach (var entry in Model.CodeCreators.OrderBy(c => c.Type).ThenBy(c => c.DisplayName))
         {
-            <div class="codecreator-panel-parameter" data-type="@entry.Type">
+            <div class="codecreator-panel-parameter">
                 <i class="@entry.IconClass"></i> @entry.DisplayName
 
                 <div class="codecreator-panel-parameter-container">
